+++ radiography/src/main/java/radiography/internal/ComposeViews.kt
@@ -1,13 +1,13 @@
 package radiography.internal
 
+import android.annotation.SuppressLint
 import android.util.SparseArray
 import android.view.View
 import androidx.compose.runtime.Composer
 import androidx.compose.runtime.Composition
-import androidx.compose.runtime.InternalComposeApi
+import androidx.compose.ui.node.RootForTest
 import androidx.compose.ui.tooling.data.UiToolingDataApi
 import androidx.compose.ui.tooling.data.asTree
-import radiography.ExperimentalRadiographyComposeApi
 import radiography.ScannableView
 import radiography.ScannableView.ChildRenderingError
 import radiography.ScannableView.ComposeView
@@ -63,9 +63,10 @@ private val viewKeyedTagsField: Field? by lazy(PUBLICATION) {
  * the right Compose artifacts aren't on the classpath, or the runtime Compose version is
  * unsupported, this function will return a [ChildRenderingError] and false.
  */
-@OptIn(ExperimentalRadiographyComposeApi::class)
+@SuppressLint("VisibleForTests")
 internal fun getComposeScannableViews(composeView: View): Pair<List<ScannableView>, Boolean> {
   var linkageError: LinkageError? = null
+
   val scannableViews = try {
     tryGetLayoutInfos(composeView)
       ?.map(ComposeLayoutInfo::toScannableView)
@@ -125,9 +126,12 @@ private fun tryGetLayoutInfos(composeView: View): Sequence<ComposeLayoutInfo>? {
   // public (eg LayoutNode), so we'd need to use even more (brittle) reflection to do that parsing.
   // That said, once Compose is more stable, it might be worth it to read the slot table directly,
   // since then we could drop the requirement for the Tooling library to be on the classpath.
-  @OptIn(InternalComposeApi::class, UiToolingDataApi::class)
+  @OptIn(UiToolingDataApi::class)
   val rootGroup = composer.compositionData.asTree()
-  return rootGroup.layoutInfos
+
+  val semanticsOwner = (composeView as? RootForTest)?.semanticsOwner
+  @OptIn(UiToolingDataApi::class)
+  return rootGroup.computeLayoutInfos(semanticsOwner = semanticsOwner)
 }
 
 private fun View.getKeyedTags(): SparseArray<*> {